+++ app/src/androidTest/java/woowacourse/movie/screening/ReservationActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.screening
 
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onData
@@ -16,6 +16,7 @@ import org.hamcrest.core.Is.`is`
 import org.hamcrest.core.IsInstanceOf.instanceOf
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.movie.R
 import woowacourse.movie.domain.screening.Movie
 import woowacourse.movie.domain.screening.Screening
 import woowacourse.movie.view.screening.ReservationActivity
