+++ app/src/main/java/woowacourse/movie/contract/ticket/TicketContract.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.contract.ticket
+
+interface TicketContract {
+    interface Presenter
+
+    interface View
+}
