+++ app/src/main/java/woowacourse/movie/presenter/ticket/TicketPresenter.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.presenter.ticket
+
+import woowacourse.movie.contract.ticket.TicketContract
+
+class TicketPresenter : TicketContract.Presenter
