+++ app/src/test/java/woowacourse/movie/ticket/TicketPresenterTest.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.ticket
+
+class TicketPresenterTest
